# DM/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /departments:
        get:
            tags:
                - DepartmentService
            operationId: DepartmentService_GetAllDepartment
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetAllDepartmentRespone'
        post:
            tags:
                - DepartmentService
            operationId: DepartmentService_CreateDepartment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.CreateDepartmentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.CreateDepartmentRespone'
    /departments/page:
        get:
            tags:
                - DepartmentService
            operationId: DepartmentService_GetDepartmentByPage
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetDepartmentByPageRespone'
    /departments/{id}:
        get:
            tags:
                - DepartmentService
            operationId: DepartmentService_GetDepartmentByID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetDepartmentByIDRespone'
        put:
            tags:
                - DepartmentService
            operationId: DepartmentService_UpdateDepartment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.UpdateDepartmentResquest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.UpdateDepartmentRespone'
        delete:
            tags:
                - DepartmentService
            operationId: DepartmentService_DeleteDepartment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.DeleteDepartmentRespone'
...
components:
    schemas:
        helloworld.v1.CreateDepartmentRequest:
            type: object
            properties:
                name:
                    type: string
        helloworld.v1.CreateDepartmentRespone:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
...
        helloworld.v1.DeleteDepartmentRespone:
            type: object
            properties: {}
...
        helloworld.v1.Department:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                subdepartments:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.SubDepartment'
        helloworld.v1.Employee:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                phone:
                    type: string
                password:
                    type: string
                salary:
                    type: integer
                    format: uint32
                role:
                    type: string
                subDepartmentId:
                    type: integer
                    format: uint32
        helloworld.v1.GetAllDepartmentRespone:
            type: object
            properties:
                departments:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.Department'
...
        helloworld.v1.GetDepartmentByIDRespone:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
        helloworld.v1.GetDepartmentByPageRespone:
            type: object
            properties:
                departments:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.Department'
...
        helloworld.v1.SubDepartment:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                departmentId:
                    type: integer
                    format: uint32
                employees:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.Employee'
        helloworld.v1.UpdateDepartmentRespone:
            type: object
            properties: {}
        helloworld.v1.UpdateDepartmentResquest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
...